% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aussiemaps-package.R
\docType{package}
\name{aussiemaps-package}
\alias{aussiemaps}
\alias{aussiemaps-package}
\title{aussiemaps: Maps of Australia}
\description{
Granular maps of Australia using ABS boundaries from 2006-2021.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://carlosyanez.github.io/aussiemaps/}
  \item Report bugs at \url{https://github.com/carlosyanez/aussiemaps/issues}
}

}
\author{
\strong{Maintainer}: Carlos Y치침ez Santib치침ez \email{carlos.yanez.s@gmail.com}

Other contributors:
\itemize{
  \item Craig Alexander \email{Craig.Alexander.2@glasgow.ac.uk} [thesis advisor]
  \item Australian Bureau of Statistics [copyright holder]
}

}
\keyword{internal}
