% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_polygons.R
\name{clean_polygons}
\alias{clean_polygons}
\title{Clean up aggregated polygons}
\usage{
clean_polygons(sfobject, tol_value = 0.05)
}
\arguments{
\item{sfobject}{sf object}

\item{tol_value}{clean up tolerance}
}
\value{
sf object
}
\description{
Clean up aggregated polygons
}
