% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{data_maps_delete}
\alias{data_maps_delete}
\title{Helper function to delete  data}
\usage{
data_maps_delete(file = NULL)
}
\arguments{
\item{file}{to delete - defaults to all of them (provide full path, can obtain from data_maps_info)}
}
\value{
nothing
}
\description{
Helper function to delete  data
}
\keyword{helpers}
