% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{data_maps_info}
\alias{data_maps_info}
\title{Helper function to update/download  data}
\usage{
data_maps_info(...)
}
\arguments{
\item{...}{fs::dir_info() parameters}
}
\value{
nothing
}
\description{
Helper function to update/download  data
}
\keyword{helpers}
