% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{find_maps_cache}
\alias{find_maps_cache}
\title{Helper function to find cache folder}
\usage{
find_maps_cache()
}
\value{
nothing
}
\description{
Helper function to find cache folder
}
\keyword{helpers}
