% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexible_left_join.R
\name{flexible_left_join}
\alias{flexible_left_join}
\title{'Flexible left join, accounting for different formatting in different datasets}
\usage{
flexible_left_join(df1, df2, by)
}
\arguments{
\item{df1}{first dataset (of sf data frame)}

\item{df2}{second dataset}

\item{by}{named vector with join key, as in left_join}
}
\value{
A data frame containing the aggregated data
}
\description{
This function's preforms for a left_join() between to datasets, accomdating for small differences in
formatting across the key, for example a O'Connor vs. Oconnor (as in the Australian electoral division)
}
