% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_aggregate.R
\name{geo_aggregate}
\alias{geo_aggregate}
\title{Aggregate data to a new geography}
\usage{
geo_aggregate(
  original_data,
  values_col,
  original_geo,
  new_geo,
  grouping_col = NULL,
  year,
  proportions_manual = NULL
)
}
\arguments{
\item{original_data}{A data frame of original data}

\item{values_col}{The name of the column containing the values to be aggregated}

\item{original_geo}{The name of the column containing the original geography}

\item{new_geo}{The name of the column containing the new geography}

\item{grouping_col}{The name of the column containing the grouping variables}

\item{year}{The year of the data to be aggregated}

\item{proportions_manual}{A data frame of manual proportions,it will override in-package proportions based on area.}
}
\value{
A data frame containing the aggregated data
}
\description{
Convert data, aggregating smaller geographic structures into larger ones.
By default it uses area to apportion values when there is no one-to-one correspondence. Weighting table can be provided
}
