% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{get_cache_name}
\alias{get_cache_name}
\title{Get name for a cached map}
\usage{
get_cache_name(
  year,
  simplification_factor,
  smoothing_threshold,
  new_crs,
  filter_table,
  aggregation
)
}
\arguments{
\item{year}{year}

\item{simplification_factor}{simplification_factor}

\item{smoothing_threshold}{smoothing_threshold}

\item{new_crs}{new_crs}

\item{filter_table}{filter_table}

\item{aggregation}{aggregation}
}
\value{
cache path
}
\description{
Get name for a cached map
}
\keyword{helpers}
