% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{Get sf object containing selected map polygons}
\usage{
get_map(
  filter_table = NULL,
  filters = NULL,
  year,
  aggregation = NULL,
  simplify = FALSE
)
}
\arguments{
\item{filter_table}{table to filter (you can start with location_table)}

\item{filters}{list with filters}

\item{year}{year}

\item{aggregation}{name of column to aggregate (POA_CODE16, LOCALITY,LGA)}

\item{simplify}{(logical) whether to simplify polygons - True by default}
}
\value{
sf object with selected polygons
}
\description{
Get sf object containing selected map polygons
}
