% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{get_region}
\alias{get_region}
\title{Obtain tibble with areas from a particular region - to be used with load_map()}
\usage{
get_region(region_name)
}
\arguments{
\item{region_name}{string with name of the region (use list_regions() to get all possible names)}
}
\value{
tibble to
}
\description{
Obtain tibble with areas from a particular region - to be used with load_map()
}
