% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_attributes}
\alias{list_attributes}
\title{List  geographic structures (or attributes)}
\usage{
list_attributes()
}
\value{
A tibble with the attributes for each year.
}
\description{
This function finds all the available geographic structures dataframe with the structures for each year.
}
\details{
Get list of all geographic structures (or attributes)
}
\examples{
\dontrun{
list_attributes()
}
}
\keyword{lists}
