% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_proportions}
\alias{list_proportions}
\title{Get list elements to attribute mapping, with area proportion}
\usage{
list_proportions(attribute_name, ids = NULL)
}
\arguments{
\item{attribute_name}{attribute name}

\item{ids}{ids}
}
\value{
tibble with structure
}
\description{
Get list elements to attribute mapping, with area proportion
}
