% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_structure}
\alias{list_structure}
\title{List Structure}
\usage{
list_structure(year, filters = NULL)
}
\arguments{
\item{year}{A character string of the year for which the structure is requested.}

\item{filters}{A list with the attributes and values of the filters to be applied  (e.g. list("CED_NAME_2021"=c("Wills","Melbourne")))}
}
\value{
A data frame with the structure requested.
}
\description{
produce table with geo structure
}
\examples{
\dontrun{
list_structure("2021")
list_structure("2021", list("CED_NAME_2021"=c("Wills","Melbourne"))
}
}
\seealso{
\code{\link{get_map}}
}
