% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_aussie_map.R
\name{load_aussie_map}
\alias{load_aussie_map}
\title{Load granular data}
\usage{
load_aussie_map(filter_table, aggregation = c("none"))
}
\arguments{
\item{filter_table}{table to filter (you can start with location_table)}

\item{aggregation}{name of column to aggregate (POA_CODE16, LOCALITY,LGA)}
}
\value{
no output
}
\description{
Load granular data
}
