% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_map.R
\name{load_map}
\alias{load_map}
\title{Load granular data}
\usage{
load_map(
  filter_table,
  aggregation = c("none"),
  clean_tolerance = 0.05,
  simplify = TRUE
)
}
\arguments{
\item{filter_table}{table to filter (you can start with location_table)}

\item{aggregation}{name of column to aggregate (POA_CODE16, LOCALITY,LGA)}

\item{clean_tolerance}{clean up tolerance}

\item{simplify}{whether to simplify (True, default) or not (False) polygon}
}
\value{
no output
}
\description{
Load granular data
}
