% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.names}
\alias{state.names}
\title{List of states and their acronym}
\format{
A data frame with 9 variables:
\describe{
\item{State_short}{2/3 letter acronym}
\item{State}{Name}
}
}
\usage{
state.names
}
\description{
List of states and their acronym
}
\keyword{datasets}
