% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{data_maps_delete}
\alias{data_maps_delete}
\title{Helper function to delete  data}
\usage{
data_maps_delete(...)
}
\arguments{
\item{...}{fs::dir_info() parameters}
}
\value{
nothing
}
\description{
Helper function to delete  data
}
\keyword{helpers}
