% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_aggregate.R
\name{geo_aggregate}
\alias{geo_aggregate}
\title{Convert data, aggregating smallar geographic structures into larger ones.}
\usage{
geo_aggregate(
  original_data,
  values_col,
  original_geo,
  new_geo,
  grouping_col,
  year,
  proportions_manual = NULL
)
}
\arguments{
\item{original_data}{data frame with data}

\item{values_col}{name of the column with numeric data}

\item{original_geo}{name of the column representing the smallest geo unit (using the names from geo structure)}

\item{new_geo}{name of the column representing the largest geo unit (using the names from geo structure)}

\item{grouping_col}{vector with all other columns to be preserved during the transformation}

\item{year}{year the geo structures belong to}

\item{proportions_manual}{manual list of correspondences, it will override in-package proportions based on area.}
}
\value{
tibble
}
\description{
Convert data, aggregating smallar geographic structures into larger ones.
}
