% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_map.R
\name{get_map}
\alias{get_map}
\title{Load sf polygons from selected locations}
\usage{
get_map(
  filter_table,
  aggregation = c("none"),
  clean_tolerance = 0.05,
  simplify = TRUE
)
}
\arguments{
\item{filter_table}{table to filter (you can start with location_table)}

\item{aggregation}{name of column to aggregate (POA_CODE16, LOCALITY,LGA)}

\item{clean_tolerance}{clean up tolerance}

\item{simplify}{(logical) whether to simplify polygons - True by default}
}
\value{
sf object with selected polygons
}
\description{
Load sf polygons from selected locations
}
