% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{Get sf object containing selected map polygons}
\usage{
get_map(
  filter_table = NULL,
  filters = NULL,
  year,
  aggregation = NULL,
  simplification_factor = 1,
  smoothing_threshold = 4
)
}
\arguments{
\item{filter_table}{table to filter (you can start with location_table)}

\item{filters}{list with filters}

\item{year}{year}

\item{aggregation}{name of column to aggregate (POA_CODE16, LOCALITY,LGA)}

\item{simplification_factor}{0-1 simplication threshold to pass to rmapshaper::ms_simplify()}

\item{smoothing_threshold}{smoothing threshold (default to 1, as in km^2)}
}
\value{
sf object with selected polygons
}
\description{
Get sf object containing selected map polygons
}
