% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_structure}
\alias{list_structure}
\title{Get list of all aggregation options}
\usage{
list_structure(year, filters = NULL)
}
\arguments{
\item{year}{year when the boundaries were releaseed (2006,2011,2016,2022)}

\item{filters}{list containing data filters (e.g. list("CED_NAME_2021"=c("Wills","Melbourne")))}
}
\value{
tibble with structure
}
\description{
Get list of all aggregation options
}
